requirements.txt
azureml-core==1.34.0
azureml-dataprep==2.22.2
azureml-train==1.34.0
torch==1.10.0
torchvision==0.11.1
pandas==1.3.3
opencv-python==4.5.3.56